fix: resolve lint issues in software SPIR-V interpreter
- unconvert: remove unnecessary type conversions (Uint32/Int32/Float32
  are type aliases, not distinct types)
- revive: rename BuiltInGlobalInvocationId -> BuiltInGlobalInvocationID,
  BuiltInLocalInvocationId -> BuiltInLocalInvocationID,
  BuiltInWorkgroupId -> BuiltInWorkgroupID,
  BuiltInSampleId -> BuiltInSampleID
- unparam: remove always-zero lo parameter from clampInt
- unused: remove unused buildLoopSumSPIRV function
- nestif: add nolint:nestif with justification for debug context checks
- maintidx: add nolint:maintidx on executeGLSLExtInst
- goconst: use existing constants (glslExtSetName, opNameLoad, etc.)
  in debug.go and interpreter.go
- staticcheck: fix ineffective break in bufferAccessChain default case
- gocritic: use switch instead of if-else in vectorShuffle,
  coord -= expr and coord++ instead of coord = coord - expr
- whitespace: remove leading newline in DispatchCompute

+++ hal/software/shader/coverage_boost_test.go
@@ -345,12 +345,12 @@ func TestAtomicMinMaxOps(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			ptr := &Pointer{Value: Uint32(tt.initial)}
+			ptr := &Pointer{Value: tt.initial}
 			interp := &interpreter{
 				module: m,
 				values: map[uint32]Value{
 					1: ptr,
-					2: Uint32(tt.operand),
+					2: tt.operand,
 				},
 			}
 			inst := Instruction{
